requirements.txt
numpy
pillow
scikit-learn
torch